move course and avgGrade code to django apps
