+++ templates/avg_grade.html
@@ -30,7 +30,7 @@
                         Opcje
                     </td>
                 </tr>
-                {% for course in courses %}
+                {% for course in all_courses %}
                     <tr>
                         <td>
                             {{ course.name }} {{ course.type }}
